Droth 1696 hazmat prohibitions traffic signs update v3 (#1398)
* DROTH-1729 Batch: create new batch to create traffic signs and support additional panels

* DROTH-1729 DataFixture batch for linear hazmat to traffic signs

* DROTH-1729 DataFixture batch for linear hazmat to traffic signs

* DROTH-1729 DataFixture batch for linear hazmat to traffic signs

* DROTH-1729 DataFixture batch for linear hazmat to traffic signs

* DROTH-1729 DataFixture batch for linear hazmat to traffic signs

* DROTH-1729 DataFixture batch for linear hazmat to traffic signs

* Cherry Pick from DROTH-1707 Improving the handling of additional signs
- Fix problems with actors when the creation of traffic signs.

* DROTH-1648 Vehicle restrictions and traffic signs
- Upgrade mergeValuesExistingOnSameRoadLink method on fillTopology method.

* DROTH-1730 Restriction Service: process to create vehicle specific restriction asset through the trafficSign adapted to additional panels

* DROTH-1731 Restriction: adapt the existing Batch to create restriction asset through the signs

* DROTH-1648 Vehicle restrictions and traffic signs
- On prohibitionService it's added a way to allow the insert of new assets;
- Upgrade mergeValuesExistingOnSameRoadLink method on fillTopology method.

* DROTH-1648 Vehicle restrictions and traffic signs
- On prohibitionService it's added a way to allow the insert of new assets;
- Upgrade mergeValuesExistingOnSameRoadLink method on fillTopology method.

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1731 Restriction: adapt the existing Batch to create restriction asset through the signs

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1733 (US 1696) Automatic process to handle multiple prohibitions

* DROTH-1734 Visualization

* DROTH-1734 Visualization

* DROTH-1787 Update a trafficSign with change the geometry should update the matched linearAsset

* DROTH-1734 Visualization

* DROTH-1734 (US 1696) Visualization

* DROTH-1880 Adapt the existing code to the new batch process

* Create new methods for unit tests
Remove trafficSign from create methods
Adapt the existing code to the new batch process

* Change and create new unit tests

* Adapt the existing code to the new batch process

* Adapt the existing code to the new batch process

* Support update trafficSign without value

* save old value on table

* Change update for trafficSign manager
Change endRoadLink convert

* TrafficSign change update actor

* DROTH-1696 changes after merge with DROH-1814

* DROTH-1696 test corrections, wrong result expected

* DROTH-1696 generator: filter non-bidirectional roads from end segments

* DROTH-1696 segments manager correction for end road links

* municipalityApiSpec correction: additional matcher added

* municipalityApiSpec correction: additional matcher added

* municipalityApiSpec correction: additional matcher added

* DROTH-1696 changes after code review: mainly unused/commented code deleted

* DROTH-1696 add_new_table_for_traffic_manager_process deleted here and modified on 1814

* DROTH-1696 "VAK-rajoitus" update with the traffic signs
- Changes after code review.

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/middleware/TrafficSignManager.scala
@@ -1,15 +1,22 @@
 package fi.liikennevirasto.digiroad2.middleware
 
 import java.sql.SQLIntegrityConstraintViolationException
+import java.util.Properties
 
-import fi.liikennevirasto.digiroad2.service.pointasset.TrafficSignInfo
+import fi.liikennevirasto.digiroad2.service.pointasset.{TrafficSignInfo, TrafficSignService}
 import fi.liikennevirasto.digiroad2._
+import fi.liikennevirasto.digiroad2.asset.Asset.DateTimePropertyFormat
 import fi.liikennevirasto.digiroad2.asset._
 import fi.liikennevirasto.digiroad2.dao.linearasset.OracleLinearAssetDao
 import fi.liikennevirasto.digiroad2.dao.pointasset.PersistedTrafficSign
 import fi.liikennevirasto.digiroad2.oracle.OracleDatabase
-import fi.liikennevirasto.digiroad2.service.RoadLinkService
+import fi.liikennevirasto.digiroad2.service.{AdditionalInformation, RoadLinkService}
 import fi.liikennevirasto.digiroad2.service.linearasset.ManoeuvreService
+import fi.liikennevirasto.digiroad2.user.UserProvider
+import org.joda.time.DateTime
+import org.json4s
+import org.json4s.jackson.Json
+import org.json4s.{CustomSerializer, DefaultFormats, Extraction, Formats, JInt, JObject, JString}
 
 object TrafficSignManager {
   val manoeuvreRelatedSigns : Seq[TrafficSignType] =  Seq(NoLeftTurn, NoRightTurn, NoUTurn)
@@ -32,23 +39,34 @@ object TrafficSignManager {
 case class TrafficSignManager(manoeuvreService: ManoeuvreService, roadLinkService: RoadLinkService) {
   def withDynTransaction[T](f: => T): T = OracleDatabase.withDynTransaction(f)
 
+  case object LinkGeomSourceSerializer extends CustomSerializer[LinkGeomSource](format => ({
+    case JInt(lg) => LinkGeomSource.apply(lg.toInt)
+  }, {
+    case lg: LinkGeomSource => JInt(lg.value)
+  }))
+
+  case object DateTimeSerializer extends CustomSerializer[DateTime](format => ( {
+    case _ => throw new NotImplementedError("DateTime deserialization")
+  }, {
+    case d: DateTime => JString(d.toString(DateTimePropertyFormat))
+  }))
+
+  protected implicit val jsonFormats: Formats = DefaultFormats + LinkGeomSourceSerializer + DateTimeSerializer
+
   lazy val linearAssetDao: OracleLinearAssetDao = {
     new OracleLinearAssetDao(roadLinkService.vvhClient, roadLinkService)
   }
 
-    def createAssets(trafficSignInfo: TrafficSignInfo, newTransaction: Boolean = true): Unit = {
-    trafficSignInfo match {
-      case trSign if TrafficSignManager.belongsToManoeuvre(trSign.signType) =>
-        manoeuvreService.createBasedOnTrafficSign(trSign, newTransaction)
+  def createAssets(trafficSignInfo: TrafficSignInfo, newTransaction: Boolean = true): Unit = {
 
-      case trSign if TrafficSignManager.belongsToProhibition(trSign.signType) =>
-        insertTrafficSignToProcess(trSign.id, Prohibition)
+      if (TrafficSignManager.belongsToManoeuvre(trafficSignInfo.signType))
+        manoeuvreService.createBasedOnTrafficSign(trafficSignInfo, newTransaction)
 
-      case trSign if TrafficSignManager.belongsToHazmat(trSign.signType) =>
-        insertTrafficSignToProcess(trSign.id, HazmatTransportProhibition)
+      else if (TrafficSignManager.belongsToProhibition(trafficSignInfo.signType))
+        insertTrafficSignToProcess(trafficSignInfo.id, Prohibition)
 
-      case _ => None
-    }
+      else if (TrafficSignManager.belongsToHazmat(trafficSignInfo.signType))
+        insertTrafficSignToProcess(trafficSignInfo.id, HazmatTransportProhibition)
   }
 
   def deleteAssets(trafficSign: Seq[PersistedTrafficSign]): Unit = {
@@ -57,24 +75,22 @@ case class TrafficSignManager(manoeuvreService: ManoeuvreService, roadLinkServic
     trafficSign.foreach { trSign =>
       val trafficSignType = trSign.propertyData.find(p => p.publicId == "trafficSigns_type").get.values.map(_.asInstanceOf[TextPropertyValue]).head.propertyValue.toInt
 
-      trafficSignType match {
-        case signType if TrafficSignManager.belongsToManoeuvre(signType) =>
-          manoeuvreService.deleteManoeuvreFromSign(manoeuvreService.withIds(Set(trSign.id)), username)
+      if (TrafficSignManager.belongsToManoeuvre(trafficSignType))
+        manoeuvreService.deleteManoeuvreFromSign(manoeuvreService.withIds(Set(trSign.id)), username)
 
-        case signType  if TrafficSignManager.belongsToProhibition(signType) =>
-            insertTrafficSignToProcess(trSign.id, Prohibition)
+       else if (TrafficSignManager.belongsToProhibition(trafficSignType))
+        insertTrafficSignToProcess(trSign.id, Prohibition, Some(trSign))
 
-        case signType if TrafficSignManager.belongsToHazmat(signType) =>
-            insertTrafficSignToProcess(trSign.id, HazmatTransportProhibition)
+      else if(TrafficSignManager.belongsToHazmat(trafficSignType))
+        insertTrafficSignToProcess(trSign.id, HazmatTransportProhibition, Some(trSign))
 
-      }
     }
   }
 
-  def insertTrafficSignToProcess(id: Long, assetInfo: AssetTypeInfo) : Unit = {
+  def insertTrafficSignToProcess(id: Long, assetInfo: AssetTypeInfo, persistedTrafficSign: Option[PersistedTrafficSign] = None) : Unit = {
     try {
       withDynTransaction {
-        linearAssetDao.insertTrafficSignsToProcess(id, assetInfo.typeId)
+        linearAssetDao.insertTrafficSignsToProcess(id, assetInfo.typeId, Json(jsonFormats).write(persistedTrafficSign.getOrElse("")))
       }
     } catch {
       case ex: SQLIntegrityConstraintViolationException => print("try insert duplicate key")
